requirements.txt
aiohappyeyeballs             2.6.1
aiohttp                      3.12.15
aiosignal                    1.4.0
annotated-types              0.7.0
anyio                        4.10.0
async-timeout                5.0.1
attrs                        25.3.0
boto3                        1.40.50
botocore                     1.40.50
cachetools                   5.5.2
certifi                      2025.8.3
charset-normalizer           3.4.3
click                        8.2.1
colorama                     0.4.6
distro                       1.9.0
exceptiongroup               1.3.0
fastapi                      0.116.1
frozenlist                   1.7.0
google-ai-generativelanguage 0.6.15
google-api-core              2.25.1
google-api-python-client     2.181.0
google-auth                  2.40.3
google-auth-httplib2         0.2.0
google-generativeai          0.8.5
googleapis-common-protos     1.70.0
groq                         0.31.1
grpcio                       1.74.0
grpcio-status                1.71.2
h11                          0.16.0
httpcore                     1.0.9
httplib2                     0.30.0
httpx                        0.28.1
idna                         3.10
jmespath                     1.0.1
joblib                       1.5.2
lxml                         6.0.1
multidict                    6.6.4
numpy                        2.2.6
packaging                    24.2
pandas                       2.3.2
pdf2image                    1.17.0
pillow                       11.3.0
pinecone                     7.3.0
pinecone-plugin-assistant    1.8.0
pinecone-plugin-interface    0.0.7
pip                          25.2
propcache                    0.3.2
proto-plus                   1.26.1
protobuf                     5.29.5
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pydantic                     2.11.7
pydantic_core                2.33.2
PyMuPDF                      1.26.5
pyparsing                    3.2.3
pypdf                        6.0.0
python-dateutil              2.9.0.post0
python-docx                  1.2.0
python-dotenv                1.1.1
python-multipart             0.0.20
# pytz                         2025.2
requests                     2.32.5
rsa                          4.9.1
s3transfer                   0.14.0
scikit-learn                 1.7.2
scipy                        1.15.3
setuptools                   80.9.0
six                          1.17.0
sniffio                      1.3.1
starlette                    0.47.3
threadpoolctl                3.6.0
tqdm                         4.67.1
typing_extensions            4.15.0
typing-inspection            0.4.1
tzdata                       2025.2
uritemplate                  4.2.0
urllib3                      2.5.0
uvicorn                      0.35.0
websocket-client             1.9.0
websockets                   13.1
wheel                        0.45.1
yarl                         1.20.1




